POC close conversation API updated with SDK 2.3.x
iOS v2.3.1
Android v.2.3.0

API works to close conversation screen on both platforms

//todo//

iOS relies on subscribing to log errors events to detect connection/jwt errors as delegate callbacks are not being triggered.
This bubbles up to JS to call the new API

need more testing to figure out why this is not working as expected.